pyproject.toml
[project]
name = "thy-name"
dependencies = [
  "alembic",
  "psycopg2-binary",
  "pydantic",
  "python-fasthtml",
  "ruff",
  "sqlalchemy",
]
